requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
yt-dlp>=2024.12.0
gdown>=5.0.0
opencv-python>=4.8.0
numpy>=1.24.0
scikit-image>=0.21.0
imagehash>=4.3.1
Pillow>=10.0.0
tqdm>=4.66.0
openai>=1.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0
moviepy>=1.0.3
python-docx>=1.0.0